# openapi_billiard.yaml
openapi: 3.0.3
info:
  title: Billiard Club Management System API
  description: Comprehensive API for managing billiard clubs, tables, employees, and billing
  version: 1.0.0
  contact:
    name: Support Team
    email: support@billardmanagement.com
  license:
    name: MIT
servers:
  - url: https://api.billardmanagement.com/v1
    description: Production Server
  - url: http://localhost:3000/v1
    description: Development Server
security:
  - bearerAuth: []
paths:
...
  /employees/{employeeId}:
    get:
      summary: Get employee details
      tags:
        - Employees
      parameters:
        - in: path
          name: employeeId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'

    patch:
      summary: Update employee information
      tags:
        - Employees
      parameters:
        - in: path
          name: employeeId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdate'
      responses:
        '200':
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'

  # ==================== PRODUCTS ====================
...
  /employees/{employeeId}/shifts:
    post:
      summary: Create employee shift
      tags:
        - Shifts
      parameters:
        - in: path
          name: employeeId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftCreate'
      responses:
        '201':
          description: Shift created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'

    get:
      summary: List shifts for an employee
      tags:
        - Shifts
      parameters:
        - in: path
          name: employeeId
          required: true
          schema:
            type: integer
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shift'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EmployeeUpdate:
      type: object
      properties:
        employee_name:
          type: string
        phone_number:
          type: string
        email:
          type: string
        salary:
          type: number
          format: decimal
        hourly_rate:
          type: number
          format: decimal
        is_active:
          type: boolean
    Employee:
      type: object
      properties:
        employee_id:
          type: integer
        club_id:
          type: integer
        employee_name:
          type: string
        employee_type:
          type: string
        phone_number:
          type: string
        email:
          type: string
        address:
          type: string
        salary:
          type: number
          format: decimal
        hourly_rate:
          type: number
          format: decimal
        date_hired:
          type: string
          format: date-time
        is_active:
          type: boolean
...
    ShiftCreate:
      type: object
      required:
        - shift_date
        - shift_type
      properties:
        shift_date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        shift_type:
          type: string
          enum: [Sáng, Chiều, Tối, Đêm]
...
    Shift:
      type: object
      properties:
        shift_id:
          type: integer
        employee_id:
          type: integer
        shift_date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        hours_worked:
          type: number
          format: decimal
        shift_type:
          type: string
        status:
          type: string
...